Move CuFileStub definition to header
